+++ src/Controllers/Teacher/InspectTeacherController.java
@@ -1,24 +1,11 @@
 package Controllers.Teacher;
 
-import Models.CourseStructure.CompulsoryModule;
-import Models.CourseStructure.CompusloryModuleConstraint;
-import Models.DatabaseBehaviours.DBController;
 import Models.Graduation.GradeAttainmentConstraint;
 import Models.Graduation.LevelOfStudyConstraint;
-import Models.Tables.Registrar.InspectRegTableRow;
 import Models.UserAccounts.Student.InsufficientCreditEnrollment;
 import Models.UserAccounts.Student.InsufficientGradeAttainment;
 import Models.UserAccounts.Student.*;;
-import Views.Registrar.InspectRegistration;
 import Views.Teacher.TeacherInspect;
-
-import java.awt.*;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.ResultSet;
-import java.sql.Statement;
-import java.util.ArrayList;
-import java.util.List;
 
 public class InspectTeacherController{
     private TeacherInspect inspectionFrame;
