+++ src/Controllers/Teacher/TeacherWelcomeScreenController.java
@@ -1,13 +1,9 @@
 package Controllers.Teacher;
 
-import Controllers.Registrar.InspectRegistrationController;
-import Controllers.Teacher.InspectTeacherController;
-import Controllers.Registrar.RegistrarWelcomeScreenController;
-import Models.Tables.Registrar.RegistrarTableRow;
 import Models.Tables.StudentGrade;
 import Models.UserAccounts.Student.Student;
-import Models.UserAccounts.Teacher;
-import Models.UserAccounts.UserAccountBuilder;
+import Models.UserAccounts.Employee.Teacher;
+import Models.UserAccounts.User.UserAccountBuilder;
 import Views.Registrar.RegistrarWelcomeScreen;
 
 public class TeacherWelcomeScreenController {
