+++ src/Models/CourseStructure/Module/CompulsoryModule.java
@@ -1,5 +1,7 @@
-package Models.CourseStructure;
+package Models.CourseStructure.Module;
 
+import Models.CourseStructure.CourseStructure;
+import Models.CourseStructure.LevelOfStudy;
 import Models.DatabaseBehaviours.DBController;
 
 import java.sql.*;
