+++ src/Models/Tables/CourseStructure/ModulesTableModel.java
@@ -1,6 +1,6 @@
 package Models.Tables.CourseStructure;
 
-import Models.CourseStructure.UniModule;
+import Models.CourseStructure.Module.UniModule;
 
 public class ModulesTableModel extends CourseStructureTableModel {
