+++ test/Models/UserAccounts/AdministratorTest.java
@@ -1,6 +1,7 @@
 package Models.UserAccounts;
 
 import Models.CourseStructure.Qualification;
+import Models.UserAccounts.Employee.Administrator;
 import Models.UserAccounts.Student.*;;
 
 import org.junit.jupiter.api.BeforeEach;
