+++ test/Models/UserAccounts/UserAccountBuilderTest.java
@@ -1,6 +1,8 @@
 package Models.UserAccounts;
 
+import Models.UserAccounts.Employee.Teacher;
 import Models.UserAccounts.Student.*;;
+import Models.UserAccounts.User.UserAccountBuilder;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
